Move TestIsARequestWithLegacyEvent under its correct package
